QR code validation logic improvement
