+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/service/submission/SubmissionService.kt
@@ -65,20 +65,8 @@ object SubmissionService {
     }
 
     fun containsValidGUID(scanResult: String): Boolean {
-        if (scanResult.length > SubmissionConstants.MAX_QR_CODE_LENGTH ||
-            scanResult.count { it == SubmissionConstants.GUID_SEPARATOR } != 1 ||
-            !matchGUIDPattern(scanResult)
-        )
-            return false
-
-        val potentialGUID = extractGUID(scanResult)
-
-        return !(potentialGUID.isEmpty() || potentialGUID.length > SubmissionConstants.MAX_GUID_LENGTH)
-    }
-
-    private fun matchGUIDPattern(scanResult: String): Boolean {
-        val pattern = SubmissionConstants.QR_CODE_REGEX
-        return pattern.toRegex().matches(scanResult)
+        val scanResult = ScanResult(scanResult)
+        return scanResult.isValid
     }
 
     fun extractGUID(scanResult: String): String =
